MultiAgent/_old_backup/duplicate_projects/chatbot_project/requirements.txt
chainlit
langchain
langchain-google-genai
langchain-groq
python-dotenv
pydantic